fix 1.7.10 baby zombie pigmen

+++ modules/v1_7_R4/src/main/java/de/Keyle/MyPet/compat/v1_7_R4/entity/types/EntityMyPigZombie.java
@@ -124,7 +124,7 @@ public class EntityMyPigZombie extends EntityMyPet {
 
     @Override
     public void updateVisuals() {
-        this.datawatcher.watch(12, getMyPet().isBaby() ? 1 : 0);
+        this.datawatcher.watch(12, (byte) (getMyPet().isBaby() ? 1 : 0));
 
         Bukkit.getScheduler().runTaskLater(MyPetApi.getPlugin(), new Runnable() {
             public void run() {
